exchange Long -> int
Long타입이 프로젝트 내에서 변환하기 힘들거같다는 의견을 수렴하여 int로 변경

+++ src/main/java/com/bringup/company/user/repository/SalaryRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface SalaryRepository extends JpaRepository<Salary, Long> {
-    Optional<Salary> findByCompanyId(Long companyId);
+public interface SalaryRepository extends JpaRepository<Salary, Integer> {
+    Optional<Salary> findByCompanyId(int companyId);
 
-    Optional<Salary> findByCompanyIdAndPosition(Long companyId, String position);
+    Optional<Salary> findByCompanyIdAndPosition(int companyId, String position);
 }
