+++ src/main/java/com/bringup/company/user/service/CompanyService.java
@@ -134,7 +134,7 @@ public class CompanyService {
         }
     }
 
-    private void saveSalary(Long companyId, SalaryDto salaryDto) {
+    private void saveSalary(int companyId, SalaryDto salaryDto) {
         Salary salary = new Salary();
         salary.setCompanyId(companyId);
         salary.setPosition(salaryDto.getPosition());
